Admin Page: Custom module override UI for traffic section
Some modules have a much different UI than other modules. For example, SEO tools and Google Analytics send the user to WordPress.com to manage the module. Verification tools shows a banner to the user to activate the module. Site stats shows a different banner to the user to activate.

This commit ensure that when the modules are inactive, we show the same style of banner for section on the traffic tab.